Move free-floating functions in views/core.py to views/utils.py
