+++ superset/views/core.py
@@ -20,20 +20,17 @@ import re
 from collections import defaultdict
 from contextlib import closing
 from datetime import datetime, timedelta
-from typing import Any, Callable, cast, Dict, List, Optional, Union
+from typing import Any, cast, Dict, List, Optional, Union
 from urllib import parse
 
 import backoff
-import msgpack
 import pandas as pd
-import pyarrow as pa
 import simplejson as json
 from flask import abort, flash, g, Markup, redirect, render_template, request, Response
 from flask_appbuilder import expose
 from flask_appbuilder.models.sqla.interface import SQLAInterface
 from flask_appbuilder.security.decorators import has_access, has_access_api
 from flask_appbuilder.security.sqla import models as ab_models
-from flask_appbuilder.security.sqla.models import User
 from flask_babel import gettext as __, lazy_gettext as _
 from sqlalchemy import and_, or_, select
 from sqlalchemy.engine.url import make_url
@@ -51,12 +48,10 @@ from superset import (
     app,
     appbuilder,
     conf,
-    dataframe,
     db,
     event_logger,
     get_feature_flags,
     is_feature_enabled,
-    result_set,
     results_backend,
     results_backend_use_msgpack,
     security_manager,
@@ -72,16 +67,13 @@ from superset.connectors.sqla.models import (
     TableColumn,
 )
 from superset.constants import RouteMethod
-from superset.errors import ErrorLevel, SupersetError, SupersetErrorType
 from superset.exceptions import (
     CertificateException,
     DatabaseNotFound,
     SupersetException,
-    SupersetSecurityException,
     SupersetTimeoutException,
 )
 from superset.jinja_context import get_template_processor
-from superset.models.core import Database
 from superset.models.dashboard import Dashboard
 from superset.models.datasource_access_request import DatasourceAccessRequest
 from superset.models.slice import Slice
@@ -97,9 +89,16 @@ from superset.typing import FlaskResponse
 from superset.utils import core as utils, dashboard_import_export
 from superset.utils.dashboard_filter_scopes_converter import copy_filter_scopes
 from superset.utils.dates import now_as_float
-from superset.utils.decorators import etag_cache, stats_timing
+from superset.utils.decorators import etag_cache
 from superset.views.database.filters import DatabaseFilter
-from superset.views.utils import get_dashboard_extra_filters
+from superset.views.utils import (
+    get_dashboard_extra_filters,
+    is_owner,
+    check_datasource_perms,
+    check_slice_perms,
+    _deserialize_results_payload,
+    get_cta_schema_name,
+)
 from superset.viz import BaseViz
 
 from .base import (
@@ -153,126 +152,6 @@ DATABASE_KEYS = [
 
 
 DATASOURCE_MISSING_ERR = __("The data source seems to have been deleted")
-USER_MISSING_ERR = __("The user seems to have been deleted")
-
-
-def is_owner(obj: Union[Dashboard, Slice], user: User) -> bool:
-    """ Check if user is owner of the slice """
-    return obj and user in obj.owners
-
-
-def check_datasource_perms(
-    self: "Superset",
-    datasource_type: Optional[str] = None,
-    datasource_id: Optional[int] = None,
-) -> None:
-    """
-    Check if user can access a cached response from explore_json.
-
-    This function takes `self` since it must have the same signature as the
-    the decorated method.
-
-    :param datasource_type: The datasource type, i.e., 'druid' or 'table'
-    :param datasource_id: The datasource ID
-    :raises SupersetSecurityException: If the user cannot access the resource
-    """
-
-    form_data = get_form_data()[0]
-
-    try:
-        datasource_id, datasource_type = get_datasource_info(
-            datasource_id, datasource_type, form_data
-        )
-    except SupersetException as ex:
-        raise SupersetSecurityException(
-            SupersetError(
-                error_type=SupersetErrorType.FAILED_FETCHING_DATASOURCE_INFO_ERROR,
-                level=ErrorLevel.ERROR,
-                message=str(ex),
-            )
-        )
-
-    if datasource_type is None:
-        raise SupersetSecurityException(
-            SupersetError(
-                error_type=SupersetErrorType.UNKNOWN_DATASOURCE_TYPE_ERROR,
-                level=ErrorLevel.ERROR,
-                message="Could not determine datasource type",
-            )
-        )
-
-    viz_obj = get_viz(
-        datasource_type=datasource_type,
-        datasource_id=datasource_id,
-        form_data=form_data,
-        force=False,
-    )
-
-    security_manager.assert_viz_permission(viz_obj)
-
-
-def check_slice_perms(self: "Superset", slice_id: int) -> None:
-    """
-    Check if user can access a cached response from slice_json.
-
-    This function takes `self` since it must have the same signature as the
-    the decorated method.
-    """
-
-    form_data, slc = get_form_data(slice_id, use_slice_data=True)
-
-    if slc:
-        viz_obj = get_viz(
-            datasource_type=slc.datasource.type,
-            datasource_id=slc.datasource.id,
-            form_data=form_data,
-            force=False,
-        )
-
-        security_manager.assert_viz_permission(viz_obj)
-
-
-def _deserialize_results_payload(
-    payload: Union[bytes, str], query: Query, use_msgpack: Optional[bool] = False
-) -> Dict[str, Any]:
-    logger.debug(f"Deserializing from msgpack: {use_msgpack}")
-    if use_msgpack:
-        with stats_timing(
-            "sqllab.query.results_backend_msgpack_deserialize", stats_logger
-        ):
-            ds_payload = msgpack.loads(payload, raw=False)
-
-        with stats_timing("sqllab.query.results_backend_pa_deserialize", stats_logger):
-            pa_table = pa.deserialize(ds_payload["data"])
-
-        df = result_set.SupersetResultSet.convert_table_to_df(pa_table)
-        ds_payload["data"] = dataframe.df_to_records(df) or []
-
-        db_engine_spec = query.database.db_engine_spec
-        all_columns, data, expanded_columns = db_engine_spec.expand_data(
-            ds_payload["selected_columns"], ds_payload["data"]
-        )
-        ds_payload.update(
-            {"data": data, "columns": all_columns, "expanded_columns": expanded_columns}
-        )
-
-        return ds_payload
-    else:
-        with stats_timing(
-            "sqllab.query.results_backend_json_deserialize", stats_logger
-        ):
-            return json.loads(payload)
-
-
-def get_cta_schema_name(
-    database: Database, user: ab_models.User, schema: str, sql: str
-) -> Optional[str]:
-    func: Optional[Callable[[Database, ab_models.User, str, str], str]] = config[
-        "SQLLAB_CTAS_SCHEMA_NAME_FUNC"
-    ]
-    if not func:
-        return None
-    return func(database, user, schema, sql)
 
 
 class AccessRequestsModelView(SupersetModelView, DeleteMixin):
